lambda/requirements.txt
requests
requests_aws4auth
elasticsearch
faker
faker_vehicle
boto3
ShopifyAPI
